LPS-184766 SF, rename HttpTuple to HttpServiceBag and move
