+++ modules/apps/static/portal-osgi-web/portal-osgi-web-http-servlet-impl/src/main/java/com/liferay/portal/osgi/web/http/servlet/internal/HttpServiceBag.java
@@ -1,27 +1,29 @@
-/*******************************************************************************
- * Copyright (c) 2014, 2015 Raymond Augé and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
  *
- * Contributors:
- *     Raymond Augé <raymond.auge@liferay.com> - Bug 436698
- ******************************************************************************/
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
 
-package com.liferay.portal.osgi.web.http.servlet.internal.util;
+package com.liferay.portal.osgi.web.http.servlet.internal;
 
-import com.liferay.portal.osgi.web.http.servlet.internal.HttpServiceRuntimeImpl;
 import com.liferay.portal.osgi.web.http.servlet.internal.servlet.ProxyServlet;
 
 import org.osgi.framework.ServiceRegistration;
 
 /**
- * @author Raymond Augé
+ * @author Dante Wang
  */
-public class HttpTuple {
+public class HttpServiceBag {
 
-	public HttpTuple(
+	public HttpServiceBag(
 		ProxyServlet proxyServlet,
 		HttpServiceRuntimeImpl httpServiceRuntimeImpl,
 		ServiceRegistration<?>... serviceRegistrations) {
